Edicion de información académica, routing y servicios
