+++ webApp/src/app/candidates/components/dashboard-cand/info-academica/create-info-acad/create-info-acad.component.html
@@ -78,7 +78,7 @@
 
     <div class="botones">
       <div class="button">
-        <button mat-raised-button color="secondary" (click)="cancelarCreacion()">Cancelar</button>
+        <a mat-raised-button color="secondary" (click)="cancelarCreacion()" routerLink="../">Cancelar</a>
       </div>
       <div class="button">
         <button mat-raised-button color="primary"
@@ -89,4 +89,3 @@
     </div>
   </form>
 </div>
-{{ formInfoAcademica.value | json }}
