--- webApp/src/app/candidates/services/list-info-academica.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ListInfoAcademicaService } from './list-info-academica.service';
-
-describe('ListInfoAcademicaService', () => {
-  let service: ListInfoAcademicaService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ListInfoAcademicaService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
